Add modman support for Mandrill and fix some dependency issues
Move the autoresponder observer from Mandrill module to
Autoresponder module to avoid having a dependency from
Mandrill to Autoresponder

Add top nav menu to get to Mandrill settings from Newsletter
menu.  This is probably just a good idea in general, but specifically
when only installing the Mandrill module, the "monkey" tab
which it uses for it's config isn't available.

Create modman file for Mandrill module