Update visitedproducts.js
When setting the Magento JavaScript merge setting in System > Developer > JavaScript Settings > Merge JavaScript Files is set to yes the frontend gives a JavaScript error. This is because the JavaScript file preceding the visitedproducts.js isn't using an ending ; to properly end the JS file. Putting an extra ; infront of the visitedproducts.js file the issue get's fixed.

Tom